+++ app/render/passes/ShadowRenderPass.cpp
@@ -139,7 +139,7 @@ void ShadowRenderPass::registerToGraph(FrameGraphBuilder& fgb, RenderContext* rc
       vkCmdSetScissor(*exeParams.cmdBuffer, 0, 1, &scissor);
 
       // Ask render context to draw big giga buffer
-      exeParams.rc->drawGigaBufferIndirect(exeParams.cmdBuffer, exeParams.buffers[1], exeParams.rc->getCurrentMeshes().size());
+      exeParams.rc->drawGigaBufferIndirect(exeParams.cmdBuffer, exeParams.buffers[1], static_cast<uint32_t>(exeParams.rc->getCurrentMeshes().size()));
 
       // Stop dynamic rendering
       vkCmdEndRendering(*exeParams.cmdBuffer);
